Update the tests and the conftest.py
Remove some fixtures to conftest.py
Created reservation/test_appointment_model.py
Created reservation/test_schedule_model.py
Created reservation/test_typeof_job_model.py
Deleted reservation/tests.py
