+++ conftest.py
@@ -75,13 +75,6 @@ def typeOfJob(professional):
 
 
 @pytest.fixture
-def persisted_typeOfJob(typeOfJob):
-    typeOfJob.professional_id.save()
-    typeOfJob.save()
-    return typeOfJob
-
-
-@pytest.fixture
 def appointment(typeOfJob, professional, client):
     return Appointment(professional_id=professional, client_id=client, typeOfJob_id=typeOfJob,
                        start_appointment=datetime(2023, 4, 17, 12, 0, 0),
@@ -90,24 +83,8 @@ def appointment(typeOfJob, professional, client):
 
 
 @pytest.fixture
-def persisted_appointment(appointment):
-    appointment.typeOfJob_id.save()
-    appointment.professional_id.save()
-    appointment.client_id.save()
-    appointment.save()
-    return appointment
-
-
-@pytest.fixture
 def schedule(professional):
     return Schedule(professional_id=professional,
                     start_day=datetime(2023, 4, 17, 10, 0, 0),
                     end_day=datetime(2023, 4, 17, 18, 0, 0),
                     meeting_time=60)
-
-
-@pytest.fixture
-def persisted_schedule(schedule):
-    schedule.professional_id.save()
-    schedule.save()
-    return schedule
